Changes by EL FISTEUR

+++ src/ch/epfl/javelo/Q28_4.java
@@ -33,7 +33,7 @@ public final class Q28_4 {
      */
 
     public static double asDouble(int q28_4) {
-        return Math.scalb(q28_4,-4);
+        return Math.scalb((double)q28_4,-4);
     }
 
     /**
@@ -43,6 +43,6 @@ public final class Q28_4 {
      */
 
     public static float asFloat(int q28_4) {
-        return (float)Math.scalb(q28_4,-4);
+        return Math.scalb((float)q28_4,-4);
     }
 }
